ブランチで作業していた内容をマージしました。 変更点 ・関数名、型名をPythonの仕様にあわせた ・文字列を引数にとる関数でwcharをcharとして扱っていたのバグを修正 ・コンストラクタ関連でバグ修正
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@5222 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ include/Base/TVector.h
@@ -344,7 +344,7 @@ public:
 	///	ベクトルのノルム(大きさ)の2乗を返す.
 	element_type square() const { return PTM::dot(exp(),exp()); }
 	///	ベクトルのノルム(大きさ)を返す.
-	element_type norm() const { return element_type( sqrt(square()) ); }
+	element_type norm() const { return element_type( sqrt((double)square()) ); }
 	///	単位ベクトルにする．
 	void unitize(){ div(norm()); }
 	///	クリッピングする(bとeを頂点とする矩形内に収まるようにする).
